requirements.txt
Flask==2.3.3
pandas==2.0.3
opencv-python-headless==4.8.1.78
face-recognition==1.3.0
numpy==1.25.2
qrcode==7.4.2
Pillow==10.0.0
openpyxl==3.1.2
gunicorn==21.2.0
